Flood Fill Algo
In image editing softwares like Paint, when we take the brush to a pixel and click, the color of the region of that pixel is replaced with a new selected color. The problem statement to do this task is as below: 
Given a 2-Dimensional screen, location of a pixel in the screen (x, y coordinates of that pixel) and a color, your task is to replace color of the given pixel and all adjacent same colored pixels with the given color.
Note: Maximum screen size in this question is taken as 50*50.